requirements.txt
Django==5.0.3
django-mptt==0.16.0
djangorestframework==3.14.0
django-seed==0.3.1
psycopg2==2.9.9
